Calendar

Description

Copyright 2011 Netrium Ltd. All rights reserved.

Module for calendars

Calendar types

type DayType = String

Day types

type CalendarPeriod = (DateTime, DateTime, DayType)

A calendar period

type Calendar = (String, [CalendarPeriod])

Define a calendar as a set of calendar days and a name

type CalendarOffsetTime = (String, Int, DayType, (Int, Int))

Define a calendar offset time as an offset to apply according to a given calendar at a given time

For instance ("EEX Power", -3, "DB", (15, 00)) will apply an offset of -3 business days according to the EEX Power calendar and return this date at 3PM

Helper functions

numTypeDays

Arguments

:: Calendar
Reference calendar

-> DateTime
Start datetime for search range

-> DateTime
End datetime for search range

-> String
Type of day

-> Int

Returns the number of days of a given type (e.g. business days) between 2 dates for a given calendar

dayInRange

Arguments

:: CalendarPeriod

-> DateTime
Start datetime for search range

-> DateTime
End datetime for search range

-> Bool

Returns true if a given calendar period is within a given date range

calDaysOffset

Arguments

:: Calendar

-> DateTime

-> String
Type of day

-> Int
Number of days to offset

-> DateTime

Returns a date which is a given number of days of a given type earlier according to a given calendar or Nothing if there is no such date

getSubCalendarByDayType

Arguments

:: Calendar

-> String
Type of day

-> String
Name of sub-calendar

-> Calendar

Get a sub-calendar from a calendar and a given day type (e.g. EEX business day calendar)

getSubCalendarByDate

Arguments

:: Calendar

-> DateTime
Reference date

-> Char
G for greater than, L for lower than, E for equal

-> String
Name of sub-calendar

-> Calendar

Get a sub-calendar from a calendar and a given reference date (e.g. all days prior to 30042011)

businessDaysOffset :: String -> DateTime -> Int -> DateTime

Returns a date n business days prior to a given date according to a given calendar

calendarDaysOffset :: String -> DateTime -> Int -> DateTime

Returns a date n calendar days prior to a given date according to a given calendar

TODO: make this work!

Calendar instances

getCalendar :: String -> Calendar

Get a given calendar. For now, calendars are manually populated here. Ideally this will done using a service and parsed here.