Common

Description

Copyright 2011 Netrium Ltd. All rights reserved.

Module for common types and functions

Common types

type Price = Obs Double

A price is modelled as an observable double

type Volume = Obs Double

A volume is modelled as an observable double

type PriceCurve = [Price]

A price curve is a list of prices

data Market

Markets, e.g. UK Power (Electricity MWh UK), NBP Gas (Gas TH UK)

Constructors

Market Commodity Unit Location

data Product

A product is defined as a market and a delivery schedule

Constructors

Product Market Schedule

type Index = Obs Double

Index, e.g. SPX

type Schedule = [DateTime]

Basic schedule, defined as a ist of date times

type SegmentedSchedule = [[DateTime]]

Advanced schedule, defined as a list of time ranges

type AmortisationTable = [(Double, Double)]

Amortisation table, defined as a list of amortisation rates and observation rates

type DayCountConvention = Double

Day count conventions, e.g. 90/360

type Underlying = Contract

Underlying assets

Dates and times

type DateTime = UTCTime

Use the UTC format

toGregorian' :: DateTime -> (Integer, Int, Int)

Deconstruct a DateTime

date :: Integer -> Int -> Int -> DateTime

Midnight on a given day

datetime :: Integer -> Int -> Int -> Int -> Int -> DateTime

A point in time on a given day

newtype DiffDateTime

The difference between 2 date times

Constructors

DiffDateTime (DateTime -> DateTime)

(<>) :: Monoid a => a -> a -> a

Used to append hours and minutes to a date, i.e. define the time within a day

adjustDateTime :: DateTime -> DiffDateTime -> DateTime

Adjust a date time by a given date time difference

daysLater :: Int -> DiffDateTime

Return a positive time difference of a given number of days

daysEarlier :: Int -> DiffDateTime

Return a negative time difference of a given number of days

monthsEarlier :: Int -> DiffDateTime

Return a positive time difference of a given number of months

monthsLater :: Int -> DiffDateTime

Return a negative time difference of a given number of months

quartersLater :: Int -> DiffDateTime

Return a negative time difference of a given number of quarters

quartersEarlier :: Int -> DiffDateTime

Return a positive time difference of a given number of quarters

yearsEarlier :: Int -> DiffDateTime

Return a positive time difference of a given number of years

yearsLater :: Int -> DiffDateTime

Return a negative time difference of a given number of years

seasons :: Int -> DiffDateTime

Return a time difference of a given number of seasons

Not implemented yet

atTimeOfDay

Arguments

:: Int
Hours

-> Int
Minutes

-> DiffDateTime

Return the time difference between midnight and a given time

onDayOfMonth :: Int -> DiffDateTime

Return a time difference of a given number of days

inMonth :: Int -> DiffDateTime

Return a time difference of a given number of months

duration

Arguments

:: Int
Hours

-> Int
Minutes

-> Int
Seconds

-> Duration

Define a time duration, in terms of hours, minutes and seconds.

convertDouble

Arguments

:: Obs Double
Double to convert

-> Obs Double
Conversion factor

-> Obs Double

Double conversion - e.g. FX, unit conversion

Extra combinators

allOf :: [Contract] -> Contract

Combine a list of contract into a contract

orZero :: ChoiceId -> Contract -> Contract

Choice between a contract and a zero contract (used for options)

Simple contract templates

physical :: Volume -> Market -> Contract

Basic physical leg

physicalWithDuration :: Volume -> Market -> Duration -> Contract

A physical leg with a duration

financial :: Price -> Currency -> CashFlowType -> Contract

Basic financial leg

fixedPrice :: Double -> Price

A fixed price: transform a double into an observable

floatingPrice :: Obs Double -> Price

A floating price

scheduledContract :: Contract -> Schedule -> Contract

Distribute the contract over a schedule

Used to apply a settlement/delivery schedule

More complex trades

zcb

Arguments

:: DateTime
Delivery date time

-> Price
Notional amount

-> Currency
Payment currency

-> CashFlowType
Cashflow type

-> Contract

A zero coupon bond: delivery of a notional amount at a given time

chooserLeg

Arguments

:: Int
Range for observation period

-> Int
Index strike for observation period

-> Time
Coupon settlement date

-> Schedule
Dates for observation period

-> Currency
Payment currency

-> Index
Coupon rate

-> CashFlowType
Cashflow type

-> Contract

Chooser Range

The note pays a coupon based on the number of days that e.g. 6-month LIBOR sits within an e.g. [80bps] range. The range is chosen by the buyer at the beginning of each coupon period.

chooserNote :: [Contract] -> Contract

A chooser note has a set of chooserLeg

forward

Arguments

:: FeeCalc
Fees

-> Market
Market, e.g. NBP Gas

-> Volume
Notional volume

-> Price
Price

-> Currency
Payment currency

-> CashFlowType
Cashflow type

-> Schedule
Schedule for physical settlement

-> Schedule
Schedule for financial settlement

-> Contract

Forward: agreement to buy or sell an asset at a pre-agreed future point

commoditySwing

Arguments

:: FeeCalc
Fees

-> Market
Market, e.g. NBP Gas

-> Price
Price

-> Currency
Payment currency

-> CashFlowType
Cashflow type

-> (Volume, Volume, Volume)
(low, normal, high) delivery volumes

-> Int
Number of exercise times

-> DiffDateTime
How far the option is in advance

-> SegmentedSchedule
Delivery schedule

-> Contract

Commodity swing: a contract guaranteeing one of two parties periodic delivery of a certain amount of commodity (the nominated amount) on certain dates in the future, within a given delivery period, at a stipulated constant price (the strike price).

The party can also change (swing) the amount delivered from the nominated amount to a new amount on a short notice, for a limited number of times

Trade metadata

newtype CounterParty

Counterparty

Constructors

CounterParty String

newtype Trader

Trader

Constructors

Trader String

newtype Book

Book

Constructors

Book String

Types for margining, fixing, netting, settlement

type MarginingSchedule = Schedule

Margining schedule

type FixingSchedule = Schedule

Fixing schedule

type SettlementSchedule = Schedule

Settlement schedule

type SettlementVolumeVariance = Obs Double

Settlement volume variance (for actuals)

data SettlementAgreement

Settlement agreement

Constructors

SettlementAgreement Market SettlementSchedule

data MarketNettingAgreement

Market netting agreement

Constructors

MarketNettingAgreement CounterParty Market

data CrossMarketNettingAgreement

Cross-market netting agreement

Constructors

CrossMarketNettingAgreement CounterParty

Fee calculations

newtype FeeCalc

Type for fee calculations

Constructors

FeeCalc (Volume -> Price -> Currency -> Int -> Contract)

calcFee

Arguments

:: FeeCalc
Function to apply to calculate the fees

-> Volume
Contract notional volume

-> Price
Contract price

-> Currency
Contract currency

-> [a]
The schedule

-> Contract

Function to calculate fees

zeroFee :: FeeCalc

No fee

initialMarginFee :: FeeCalc

Fees for initial margin

computedFee :: CashFlowType -> Price -> FeeCalc

Basic calculated fees

fixedFee :: CashFlowType -> Double -> FeeCalc

Fixed fees

exchangeFee :: Price -> FeeCalc

Exchange fees

andFee :: FeeCalc -> FeeCalc -> FeeCalc

Append 2 fee calculations

type TradedDate = Obs DateTime

Traded date

data Direction

Contract direction (give or take)

Constructors

DGive

DTake

Utilities

getAmRate :: Obs Double -> AmortisationTable -> Obs Double

Function to determine amortisation rate from an amortisation table and an observable rate

List helper functions

lastn :: Int -> [a] -> [a]

Extract n last elements from a list

Basic math functions

square :: Num n => n -> n

Square a double