Contract

Description

Copyright 2011 Netrium Ltd. All rights reserved.

Module for contract

Contract type definition

data Tradeable

A canonical tradeable element, physical or financial

Constructors

Physical Commodity Unit Location (Maybe Duration)

Financial Currency CashFlowType

newtype Duration

A duration is a span of time, measured in seconds.

Constructors

Duration Int

newtype Commodity

Commodity, e.g. Gas, Electricity

Constructors

Commodity String

newtype Unit

Unit, e.g. tonnes, MWh

Constructors

Unit String

newtype Location

Location, e.g. UK, EU

Constructors

Location String

newtype Currency

Currency, e.g. EUR, USD, GBP

Constructors

Currency String

newtype CashFlowType

Cashflow type, e.g. cash, premium

Constructors

CashFlowType String

type ScaleFactor = Double

Scaling factor (used to scale the One contract)

type ChoiceId = String

Choice label, used for options

data Contract

Contract

Constructors

Zero
Deliver nothing now

One Tradeable
Deliver one tradeable element now

Give Contract
Give a contract

And Contract Contract
Append 2 contracts

Or ChoiceId Contract Contract
Offer a choice between 2 contracts

Cond (Obs Bool) Contract Contract
The contract returned depend on a boolean condition

Scale (Obs Double) Contract
Scale a contract using a scaling factor

Read Var (Obs Double) Contract
Read a variable into a contract

When (Obs Bool) Contract
Return a contract only at a given time

Anytime ChoiceId (Obs Bool) Contract
Return a contract anytime a condition is met

Until (Obs Bool) Contract
Return a contract until a condition is met

type Var = String

A variable

type Time = UTCTime

We use a continuous model of time.

data Obs a

Domain-specific language of observable values. Morally, an observable is a function from Time to a result.

konst :: (Show a, Eq a) => a -> Obs a

Constant value

var :: VarName -> Obs Double

primVar :: VarName -> Obs Double

Map a string to an observable (double)

primCond :: VarName -> Obs Bool

Map a string to an observable (boolean)

at :: Time -> Obs Bool

Returns true if we are at a given point in time

before :: Time -> Obs Bool

Returns true if we are before a given point in time

after :: Time -> Obs Bool

Returns true if we are after a given point in time

between :: Time -> Time -> Obs Bool

Returns true if we are after between 2 given point in time

between t1 t2 = time >= t1 && time < t2

ifthen :: Obs Bool -> Obs a -> Obs a -> Obs a

if..then..else for observables (returns an observable)

negate :: Num a => a -> a

max :: Obs Double -> Obs Double -> Obs Double

min :: Obs Double -> Obs Double -> Obs Double

abs :: Num a => a -> a

(%==) :: Obs Double -> Obs Double -> Obs Bool

(%>) :: Obs Double -> Obs Double -> Obs Bool

(%>=) :: Obs Double -> Obs Double -> Obs Bool

(%<) :: Obs Double -> Obs Double -> Obs Bool

(%<=) :: Obs Double -> Obs Double -> Obs Bool

(%&&) :: Obs Bool -> Obs Bool -> Obs Bool

(%||) :: Obs Bool -> Obs Bool -> Obs Bool

(%+) :: Obs Double -> Obs Double -> Obs Double

(%-) :: Obs Double -> Obs Double -> Obs Double

(%*) :: Obs Double -> Obs Double -> Obs Double

(%/) :: Obs Double -> Obs Double -> Obs Double