Observable

Description

Copyright 2011 Netrium Ltd. All rights reserved.

Module for observables

Documentation

type Time = UTCTime

We use a continuous model of time.

mkdate :: Integer -> Int -> Int -> Time

Convenience function to create a time from a date.

data Obs a where

Domain-specific language of observable values. Morally, an observable is a function from Time to a result.

Constructors

Const :: (Show a, Eq a) => a -> Obs a

Var :: VarName -> Obs Double

NamedVal :: VarName -> Obs Double

NamedCond :: VarName -> Obs Bool

Pi :: Obs Double

At :: Time -> Obs Bool

After :: Time -> Obs Bool

Before :: Time -> Obs Bool

UnOp :: UnOp a b -> Obs a -> Obs b

BinOp :: BinOp a a b -> Obs a -> Obs a -> Obs b

IfThen :: Obs Bool -> Obs a -> Obs a -> Obs a

konst :: (Show a, Eq a) => a -> Obs a

Constant value

type VarName = String

A variable name

var :: VarName -> Obs Double

primVar :: VarName -> Obs Double

Map a string to an observable (double)

primCond :: VarName -> Obs Bool

Map a string to an observable (boolean)

at :: Time -> Obs Bool

Returns true if we are at a given point in time

before :: Time -> Obs Bool

Returns true if we are before a given point in time

after :: Time -> Obs Bool

Returns true if we are after a given point in time

between :: Time -> Time -> Obs Bool

Returns true if we are after between 2 given point in time

between t1 t2 = time >= t1 && time < t2

(%==) :: Obs Double -> Obs Double -> Obs Bool

(%>) :: Obs Double -> Obs Double -> Obs Bool

(%>=) :: Obs Double -> Obs Double -> Obs Bool

(%<) :: Obs Double -> Obs Double -> Obs Bool

(%<=) :: Obs Double -> Obs Double -> Obs Bool

(%&&) :: Obs Bool -> Obs Bool -> Obs Bool

(%||) :: Obs Bool -> Obs Bool -> Obs Bool

(%+) :: Obs Double -> Obs Double -> Obs Double

(%-) :: Obs Double -> Obs Double -> Obs Double

(%*) :: Obs Double -> Obs Double -> Obs Double

(%/) :: Obs Double -> Obs Double -> Obs Double

ifthen :: Obs Bool -> Obs a -> Obs a -> Obs a

if..then..else for observables (returns an observable)

negate :: Num a => a -> a

not :: Obs Bool -> Obs Bool

Negate a boolean observable

max :: Obs Double -> Obs Double -> Obs Double

min :: Obs Double -> Obs Double -> Obs Double

abs :: Num a => a -> a

parseObsCond :: XMLParser (Obs Bool)

XML parser for booleans

parseObsReal :: XMLParser (Obs Double)

XML parser for doubles

printObs :: Obs a -> Content ()

Create XML tags

eval :: Time -> Obs a -> Steps a

Evaluate an observable at a given time

data Steps a

Constructors

NeedNamedVal Time VarName (Double -> Steps a)

NeedNamedCond Time VarName (Bool -> Steps a)

Result a

subst :: VarName -> Double -> Obs a -> Obs a

isTrue :: Time -> Obs Bool -> Bool

Check if an observable is true at a given point in time

isFalse :: Time -> Obs Bool -> Bool

Check if an observable is false at a given point in time

nextTrue :: Time -> Obs Bool -> Maybe Time

If an observable is true at a given point in time, return this time

nextFalse :: Time -> Obs Bool -> Maybe Time

If an observable is false at a given point in time, return this time

evermoreTrue :: Time -> Obs Bool -> Bool

evermoreFalse :: Time -> Obs Bool -> Bool

timeHorizon :: Time -> Obs Bool -> Maybe Time

See nextTrue

earliestTimeHorizon :: Time -> [(Obs Bool, a)] -> Maybe (Time, a)

Return the earliest time an observable is true (or Nothing if it is never true)

simplifyWithinHorizon :: Time -> Time -> Obs Bool -> Obs Bool