Options

Description

Copyright 2011 Netrium Ltd. All rights reserved.

Module for options

Options are built up as follows:

In general terms, an option is a contract that gives you the right, under certain conditions, to buy (call) or sell (put) some underlying asset.

There are two separate aspects to an "underlying":

There are the rights (ie a contract) you get when you choose to exercise the option (e.g. the right to buy some physical or financial asset)

An observable, usually related to the spot price of the physical or financial asset that the option involves.

Types

type ExerciseCondition = Contract -> Contract

Add an exercise condition to a contract, e.g. to add a barrier condition

type ConditionWindow = [(DateTime, DateTime)]

Define a time window when a condition applies as a set of time ranges

type ExpirationCondition = Contract -> Contract

An expiration condition

type UnderlyingPrice = Price

The spot price of the underlying asset

type UnderlyingContract = StrikePrice -> Contract

The rights to acquire the underlying asset. This must include any payments (which may depend on the option strike price).

type StrikePrice = Price

The price the option owner pays to acquire the underlying

data OptionDirection

The direction of the option (put or call)

Constructors

CallOption
Option to buy underlying

PutOption
Option to sell underlying

newtype ExerciseDetails

Option contracts differ in the detail of when and at what price the option can be exercised.

Constructors

ExerciseDetails (ChoiceId -> (StrikePrice -> Contract) -> Contract)

Option template

option

Arguments

:: ChoiceId
Choice label

-> ExerciseDetails
Details of when and at what price the option can be exercised

-> OptionDirection
Direction of the option (call or put)

-> Obs Double
Premium

-> Currency
Currency for the premium

-> Maybe Schedule
Payment schedule, if Nothing then premium is paid instantly

-> UnderlyingContract
Underlying asset

-> Contract

Basic option template. Flexibility is achieved through ExerciseDetails.

Templates for option parameters

Exercise time

europeanExercise :: DateTime -> StrikePrice -> ExerciseDetails

European exercise: option may be exercised only at the expiry date of the option, i.e. at a single pre-defined point in time.

americanExercise :: (DateTime, DateTime) -> StrikePrice -> ExerciseDetails

American exercise: option may be exercised at any time before the expiry date.

bermudanExercise :: [(DateTime, DateTime)] -> StrikePrice -> ExerciseDetails

Bermudan exercise: option may be exercised at a set (always discretely spaced) number of times.

Payoff

asianExercise :: UnderlyingPrice -> Schedule -> ExerciseDetails

Asian exercise: option where the payoff is not determined by the underlying price at maturity but by the average underlying price over some pre-set period of time.

Exercise conditions

barrierKnockIn :: Obs Bool -> ExerciseDetails -> ExerciseDetails

Barrier knock-in: options that start their lives worthless and only become active in the event a predetermined knock-in barrier price is breached Barrier options become activated or, on the contrary, null and void only if the underlier reaches a predetermined level (barrier).

barrierUpAndIn :: Index -> Price -> ExerciseDetails -> ExerciseDetails

Barrier up-and-in: spot price starts below the barrier level and has to move up for the option to become activated

barrierDownAndIn :: Index -> Price -> ExerciseDetails -> ExerciseDetails

Barrier down-and-in: spot price starts above the barrier level and has to move down for the option to become activated.

More advanced option templates

commoditySpreadOption

Arguments

:: ChoiceId
Choice label

-> [(Market, Volume, Price, Currency, CashFlowType, SegmentedSchedule, FeeCalc)]
List of underlying legs

-> CalendarOffsetTime
Exercise calendar name, offset, time of the day and type of day for exercise

-> CalendarOffsetTime
Payment calendar name, offset, time of the day and type of day for payment

-> OptionDirection
Option direction (put or call)

-> StrikePrice
Strike price of the option

-> Currency
Currency of the strike price

-> CashFlowType
Cashflow type of the strike price

-> Obs Double
Premium

-> Currency
Currency for the premium

-> Contract

Commodity Spread Option: a strip of options with a spread underlying (e.g. with x legs)

Exercise is determined by an offset to the earliest delivery date of the underlying for each given option

Generic so:

Options can be daily, monthly or any grain/combination of grains

Underlying can have any number of legs (the grain of the leg does not have to be the same)