Swaps

Description

Copyright 2011 Netrium Ltd. All rights reserved.

Module for swap contracts

Documentation

commoditySwap :: (Market, Market) -> (Volume, Volume) -> Price -> Currency -> CashFlowType -> Schedule -> Contract

Basic commodity swap: simultaneous swap of commodity A for commodity B for a nominal cost over a delivery schedule (same cost for each delivery period)

varianceSwap

Arguments

:: Price
Notional amount

-> Obs Double
Vega amount

-> Obs Double
Variance amount

-> Schedule
Observation schedule

-> Index
Placeholder for index, currently not used

-> [Price]
Official closing prices of the underlying over the observation period

-> Currency
Payment currency

-> CashFlowType
Payment cashflow type

-> DiffDateTime
Settlement offset

-> String
Calendar to use

-> Contract

Variance swap

A variance swap is an instrument which allows investors to trade future realized (or historical) volatility against current implied volatility. Only variance the squared volatility can be replicated with a static hedge.

indexAmortisingSwap

Arguments

:: Price
Notional amount

-> Schedule
Coupon dates

-> (Currency, Currency)
Currencies respectively for swap buyer and seller

-> (CashFlowType, CashFlowType)
Cashflow types respectively for cash buyer and seller

-> AmortisationTable
Amortisation table

-> DayCountConvention
Day count convention, e.g. 90/360

-> Index
Buyer rate

-> Index
Seller rate

-> Contract

Index Amortising Swap

Swap Buyer receives the following coupons: Buyer rate paid e.g. quarterly (according to specified day count convention) on the principal amount, amortised according to the values of the buyer rate

Swap Seller receives seller rate on same coupon dates.

creditDefaultSwap

Arguments

:: [CreditEvent]
List of credit events

-> Price
Notional amount

-> Schedule
Payment schedule

-> Currency
Payment currency

-> Index
Payment rate

-> CashFlowType
Payment cashflow type

-> DayCountConvention
Day count convention, e.g. 90/360

-> Underlying
Underlying asset

-> Contract

Credit Default Swap

CDS buyer pays an agreed rate on an agreed basis on the principal amount

CDS seller makes capital payments if and when credit events happen